refactor(#3113): move shared-vocabulary js-tag below the IR layer (slice 1)
The JsTag enum + jsTagUnboxKind live in a dependency-free leaf (js-tag.ts,
extracted in #2949) that is shared vocabulary consumed by IR core files
(nodes/verify/builder/from-ast) AND by codegen (value-tags.ts). It lived in
src/codegen/, so the IR files reaching for it created an IR->codegen import
inversion (the intended layering is emit <- ir <- codegen).

Move src/codegen/js-tag.ts -> src/ir/js-tag.ts (pure relocation; the file has
zero imports, no logic change). Update all consumers to the new path: the 6 IR
importers now import in-layer; codegen/value-tags.ts imports down-stack via
../ir/js-tag.js (re-export preserved); the 7 issue-2949-* tests updated. A thin
@deprecated re-export stub stays at src/codegen/js-tag.ts for one cycle so the
concurrently-edited #2855/#2856 IR-migration branches keep compiling while they
rebase.

Effect: ir/nodes.ts, ir/verify.ts, ir/builder.ts had js-tag as their ONLY
codegen import -> now ZERO codegen imports; ir/from-ast.ts and ir/integration.ts
each drop one.

Proof: scripts/prove-emit-identity.mjs check reports IDENTICAL across all 56
(file,target) emits; tsc --noEmit clean; issue-2949-* suites pass (41 tests).

Scope: this is slice 1 (Problem 1) of #3113. Slice 2 — containing the 2,610-LOC
ir/integration.ts IR->codegen bridge — is the larger, design-sensitive part and
stays open under #3113.

Pre-commit checklist reviewed. ✓
